+++ cookbooks/aws-parallelcluster-environment/metadata.rb
@@ -9,7 +9,7 @@ source_url 'https://github.com/aws/aws-parallelcluster-cookbook'
 chef_version '>= 18'
 version '3.13.0'
 
-depends 'line', '~> 4.5.13'
-depends 'nfs', '~> 5.1.2'
+depends 'line', '~> 4.5.21'
+depends 'nfs', '~> 5.1.5'
 
 depends 'aws-parallelcluster-shared', '~> 3.13.0'