+++ cookbooks/aws-parallelcluster-platform/metadata.rb
@@ -9,6 +9,6 @@ source_url 'https://github.com/aws/aws-parallelcluster-cookbook'
 chef_version '>= 18'
 version '3.13.0'
 
-depends 'line', '~> 4.5.13'
+depends 'line', '~> 4.5.21'
 
 depends 'aws-parallelcluster-shared', '~> 3.13.0'