+++ cookbooks/aws-parallelcluster-shared/metadata.rb
@@ -9,5 +9,5 @@ source_url 'https://github.com/aws/aws-parallelcluster-cookbook'
 chef_version '>= 18'
 version '3.13.0'
 
-depends 'yum', '~> 7.4.13'
-depends 'yum-epel', '~> 5.0.2'
+depends 'yum', '~> 7.4.20'
+depends 'yum-epel', '~> 5.0.8'