+++ cookbooks/third-party/openssh-2.11.14/spec/spec_helper.rb
@@ -1,0 +1,14 @@
+require 'chefspec'
+require 'chefspec/berkshelf'
+
+RSpec.configure do |config|
+  config.color = true               # Use color in STDOUT
+  config.formatter = :documentation # Use the specified formatter
+  config.log_level = :error         # Avoid deprecation notice SPAM
+  config.platform = 'ubuntu'
+  config.version = '18.04'
+end
+
+def join_path(*path)
+  Chef::Util::PathHelper.cleanpath(::File.join(path))
+end